develop 2: refactor order and user test and add new allure report
