+++ src/main/java/user/AuthorizationUserApi.java
@@ -4,9 +4,9 @@ import baseurl.BaseHttpClient;
 import io.restassured.response.Response;
 
 public class AuthorizationUserApi extends BaseHttpClient {
-    private final String apiPathCreate = "/api/auth/login";
+    private final String apiPathLogin = "/api/auth/login";
 
     public Response postAuthUser(User user) {
-        return doNotAuthPostRequest(apiPathCreate, user);
+        return doNotAuthPostRequest(apiPathLogin, user);
     }
 }
