+++ src/main/java/user/LogoutUserApi.java
@@ -3,7 +3,7 @@ package user;
 import baseurl.BaseHttpClient;
 import io.restassured.response.Response;
 
-public class LoguotUserApi extends BaseHttpClient {
+public class LogoutUserApi extends BaseHttpClient {
     private final String apiPathLogout = "/api/auth/logout";
 
     public Response postUser(String accessToken, String refreshToken) {
